+++ CHANGELOG.md
@@ -17,6 +17,7 @@ tech changes will usually be stripped from release notes for the public
 
 -   Assets not being able to moved up to parent folder
 -   Assets not being removable if a shape with a link to the asset exists
+-   Vision tool alert state was not always accurate
 
 ## [2022.2.2] - 2022-06-17